// react-module/src/openapi/models/for-pack.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ForPack
 */
export interface ForPack {
    /**
     * 
     * @type {number}
     * @memberof ForPack
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof ForPack
     */
    barcodeFromId?: any;
    /**
     * 
     * @type {string}
     * @memberof ForPack
     */
    barcodeEndProduct?: any;
    /**
     * 
     * @type {LayerShapeExtendedDto}
     * @memberof ForPack
     */
    layerShape?: any;
    /**
     * 
     * @type {AtomicProductDefinitionBasicDto}
     * @memberof ForPack
     */
    atomicProductDefinition?: any;
    /**
     * 
     * @type {number}
     * @memberof ForPack
     */
    layerNumber?: any;
    /**
     * 
     * @type {OnlyErpOrder}
     * @memberof ForPack
     */
    productionOrder?: any;
}