// react-module/src/openapi/models/scanned-barcode-verification.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ScannedBarcodeVerification
 */
export interface ScannedBarcodeVerification {
    /**
     * 
     * @type {string}
     * @memberof ScannedBarcodeVerification
     */
    barcode?: any;
    /**
     * 
     * @type {string}
     * @memberof ScannedBarcodeVerification
     */
    status?: ScannedBarcodeVerificationStatusEnum;
    /**
     * 
     * @type {number}
     * @memberof ScannedBarcodeVerification
     */
    assignedPackIfDiffersId?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum ScannedBarcodeVerificationStatusEnum {
    OK = 'OK',
    CANBEASSIGNED = 'CAN_BE_ASSIGNED',
    ASSIGNEDTOANOTHERPACK = 'ASSIGNED_TO_ANOTHER_PACK',
    DOESNOTEXIST = 'DOES_NOT_EXIST'
}